+++ src/diffusers/pipelines/t2i_adapter/pipeline_stable_diffusion_xl_adapter.py
@@ -161,18 +161,10 @@ def retrieve_timesteps(
     timesteps: Optional[List[int]] = None,
     sigmas: Optional[List[float]] = None,
     **kwargs,
-) :
+):
     r"""
-    Calls the scheduler's `set_timesteps` method and retrieves timesteps from the scheduler after the call.
-    Handles custom timesteps. Any kwargs will be supplied to `scheduler.set_timesteps`.
-
-    Backwards compatible: by default the function behaves exactly as before and returns
-        (timesteps_tensor, num_inference_steps)
-
-    If the caller passes `return_scheduler=True` in kwargs, the function will **not** mutate the passed
-    scheduler. Instead it will use a cloned scheduler if available (via `scheduler.clone_for_request`)
-    or a deepcopy fallback, call `set_timesteps` on that cloned scheduler, and return:
-        (timesteps_tensor, num_inference_steps, scheduler_in_use)
+    Calls the scheduler's `set_timesteps` method and retrieves timesteps from the scheduler after the call. Handles
+    custom timesteps. Any kwargs will be supplied to `scheduler.set_timesteps`.
 
     Args:
         scheduler (`SchedulerMixin`):
@@ -189,72 +181,36 @@ def retrieve_timesteps(
             Custom sigmas used to override the timestep spacing strategy of the scheduler. If `sigmas` is passed,
             `num_inference_steps` and `timesteps` must be `None`.
 
-    Optional kwargs:
-        return_scheduler (bool, default False): if True, return (timesteps, num_inference_steps, scheduler_in_use)
-            where `scheduler_in_use` is a scheduler instance that already has timesteps set.
-            This mode will prefer `scheduler.clone_for_request(...)` if available, to avoid mutating the original scheduler.
-
     Returns:
-        `(timesteps_tensor, num_inference_steps)` by default (backwards compatible), or
-        `(timesteps_tensor, num_inference_steps, scheduler_in_use)` if `return_scheduler=True`.
+        `Tuple[torch.Tensor, int]`: A tuple where the first element is the timestep schedule from the scheduler and the
+        second element is the number of inference steps.
     """
-    # pop our optional control kwarg (keeps compatibility)
-    return_scheduler = bool(kwargs.pop("return_scheduler", False))
-
     if timesteps is not None and sigmas is not None:
         raise ValueError("Only one of `timesteps` or `sigmas` can be passed. Please choose one to set custom values")
-
-    # choose scheduler to call set_timesteps on
-    scheduler_in_use = scheduler
-    if return_scheduler:
-        # Do not mutate the provided scheduler: prefer to clone if possible
-        if hasattr(scheduler, "clone_for_request"):
-            try:
-                # clone_for_request may accept num_inference_steps or other kwargs; be permissive
-                scheduler_in_use = scheduler.clone_for_request(num_inference_steps=num_inference_steps or 0, device=device)
-            except Exception:
-                scheduler_in_use = copy.deepcopy(scheduler)
-        else:
-            # fallback deepcopy (scheduler tends to be smallish - acceptable)
-            scheduler_in_use = copy.deepcopy(scheduler)
-
-    # helper to test if set_timesteps supports a particular kwarg
-    def _accepts(param_name: str) -> bool:
-        try:
-            return param_name in set(inspect.signature(scheduler_in_use.set_timesteps).parameters.keys())
-        except (ValueError, TypeError):
-            # if signature introspection fails, be permissive and attempt the call later
-            return False
-
-    # now call set_timesteps on the chosen scheduler_in_use (may be original or clone)
     if timesteps is not None:
-        accepts_timesteps = _accepts("timesteps")
+        accepts_timesteps = "timesteps" in set(inspect.signature(scheduler.set_timesteps).parameters.keys())
         if not accepts_timesteps:
             raise ValueError(
-                f"The current scheduler class {scheduler_in_use.__class__}'s `set_timesteps` does not support custom"
+                f"The current scheduler class {scheduler.__class__}'s `set_timesteps` does not support custom"
                 f" timestep schedules. Please check whether you are using the correct scheduler."
             )
-        scheduler_in_use.set_timesteps(timesteps=timesteps, device=device, **kwargs)
-        timesteps_out = scheduler_in_use.timesteps
-        num_inference_steps = len(timesteps_out)
+        scheduler.set_timesteps(timesteps=timesteps, device=device, **kwargs)
+        timesteps = scheduler.timesteps
+        num_inference_steps = len(timesteps)
     elif sigmas is not None:
-        accept_sigmas = _accepts("sigmas")
+        accept_sigmas = "sigmas" in set(inspect.signature(scheduler.set_timesteps).parameters.keys())
         if not accept_sigmas:
             raise ValueError(
-                f"The current scheduler class {scheduler_in_use.__class__}'s `set_timesteps` does not support custom"
+                f"The current scheduler class {scheduler.__class__}'s `set_timesteps` does not support custom"
                 f" sigmas schedules. Please check whether you are using the correct scheduler."
             )
-        scheduler_in_use.set_timesteps(sigmas=sigmas, device=device, **kwargs)
-        timesteps_out = scheduler_in_use.timesteps
-        num_inference_steps = len(timesteps_out)
+        scheduler.set_timesteps(sigmas=sigmas, device=device, **kwargs)
+        timesteps = scheduler.timesteps
+        num_inference_steps = len(timesteps)
     else:
-        # default path
-        scheduler_in_use.set_timesteps(num_inference_steps, device=device, **kwargs)
-        timesteps_out = scheduler_in_use.timesteps
-
-    if return_scheduler:
-        return timesteps_out, num_inference_steps, scheduler_in_use
-    return timesteps_out, num_inference_steps
+        scheduler.set_timesteps(num_inference_steps, device=device, **kwargs)
+        timesteps = scheduler.timesteps
+    return timesteps, num_inference_steps
 
 
 class StableDiffusionXLAdapterPipeline(
