+++ src/diffusers/schedulers/scheduling_ddpm_wuerstchen.py
@@ -161,11 +161,6 @@ class DDPMWuerstchenScheduler(SchedulerMixin, ConfigMixin):
             timesteps = torch.Tensor(timesteps).to(device)
         self.timesteps = timesteps
 
-    def clone_for_request(self, num_inference_steps: int, device: Union[str, torch.device] = None, timesteps: Optional[List[int]] = None):
-        import copy
-        local = copy.deepcopy(self)
-        local.set_timesteps(num_inference_steps=num_inference_steps, device=device, timesteps=timesteps)
-        return local
 
     def step(
         self,
